backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "gtts>=2.5.4",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.57.6",
    "uvicorn>=0.40.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
